Bug 1794163 - Fix C++20 -Wenum-float-conversion warning in resistFingerprinting. r=tjr
C++20 deprecates some operations between enums and floating point types:

https://clang.llvm.org/docs/DiagnosticsReference.html#wenum-float-conversion

toolkit/components/resistfingerprinting/nsRFPService.cpp:428:56 [-Wenum-float-conversion] arithmetic between floating-point type 'double' and enumeration type 'mozilla::nsRFPService::TimeScale'

Differential Revision: https://phabricator.services.mozilla.com/D158866

UltraBlame original commit: 77d4387553a6ae95d0cf8ebf46d804e0bb939ebc

+++ toolkit/components/resistfingerprinting/nsRFPService.cpp
@@ -425,7 +425,7 @@ double nsRFPService::ReduceTimePrecisionImpl(double aTime, TimeScale aTimeScale,
   }
 
   
-  double ret = double(clampedAndJittered) / (1000000.0 / aTimeScale);
+  double ret = double(clampedAndJittered) / (1000000.0 / double(aTimeScale));
 
   MOZ_LOG(
       gResistFingerprintingLog, LogLevel::Verbose,
